WIP: reuse graphics devices over multiple tests
This change reuses the graphics devices so that we don't keep creating
and destroying them for every test case.

This improves the runtime of slang-test by 40%. And it matches to the
initial expectation because when profiled, _createDevice was taking 40%
of the total execution time.

TODO: when the testing is done, the cached devices need to be explicitly
released in TestContext::~TestContext().