Fix issue in ordinary data writing in argument buffer (#331)
Just revert the ordinary fields data writing to previous version.
As in #303, the ordinary data writing is based on the assumption that all the shaderObject in writeArgumentBuffer call is already a argument buffer type layout. But that is not the case, so simply memcpy the data won't work.

Revert to previous version now.