Add tests that exercise ReorderThread and Invoke (#447)
* Add tests that exercise ReorderThread and Invoke

The tests check that we get correct results when we create a hit object,
reorder, then invoke the hit obj.

Note that we don't check that the reorder call actually does anything,
because whether or not the implementation chooses to reorder the threads
is unspecified.

* Enable raytracing reorder test for all APIs

To do so, fix a bug where the D3D12 implementation didn't add the UAV
slot for NVAPI to the root descriptor table.

Also run pre-commit hooks.

* Put NVAPI UAV resource descriptor in its own set

Doing so fixes a failure the previous commit introduced in the test
root-shader-parameter.

* Undo NVAPI + D3D12 root signature change

Disable the reorder test for D3D12 and undo the changes attempting to
add the NVAPI dummy UAV to the root signature.