Reduce the number of cmake cache variables, especially for configuring dependencies. (#464)
This PR removes a lot of the cache variables for slang-rhi's dependencies, in order to make it easier to switch between branches and keep dependencies automatically up-to-date.

The remaining cache variables are:
- SLANG_RHI_ENABLE_D3D11
- SLANG_RHI_ENABLE_D3D12
- SLANG_RHI_ENABLE_AGILITY_SDK
- SLANG_RHI_ENABLE_NVAPI
- SLANG_RHI_ENABLE_VULKAN
- SLANG_RHI_ENABLE_METAL
- SLANG_RHI_ENABLE_CUDA
- SLANG_RHI_ENABLE_OPTIX
- SLANG_RHI_ENABLE_WGPU
- SLANG_RHI_FETCH_SLANG
- SLANG_RHI_FETCH_SLANG_VERSION
- SLANG_RHI_SLANG_INCLUDE_DIR
- SLANG_RHI_SLANG_BINARY_DIR

All the URLs for the dependencies are now normal variables set to sensible default values and generally have the form SLANG_RHI_XXX_URL (removed the _FETCH_ part) so they don't pick up old values during the transition.