Add deferTargetCompilation to pipeline descs (#513)
* Add deferTargetCompilation to pipeline descs

This flag allows for deferred target code compilation at dispatch time. In the future, we plan to compile all target code needed for a command buffer in parallel. For now, compilation still happens single-threaded.

* Fix creation of concrete pipelines

The main change is that for non-specializable programs, we create the concrete pipeline once and store a reference to in the virtual pipeline,, tying the lifetime of the concrete pipeline to the virtual (user-facing) one and makeing later lookup fast.