Put more HitObject intrinsics under test (#539)
* Rename ray-tracing-a and ray-tracing-b tests

Give those tests more descriptive names so it's clear where they fit in
with the rest of the ray tracing tests.

* Enable reorder tests under D3D12

These tests were previously disabled because of a bug in slang-rhi, but
that bug has been fixed and they now pass under all backends.

* Add tests to exercise creating, querying, and invoking hit objects

* add tests for GetRayObjectDirection and GetRayObjectOrigin

* Extend DeviceExtraOptions so tests can limit max shader model

* Add test for MakeHit

* run commit hooks

* rename d3d12MaxShaderModel to d3d12HighestShaderModel

* Add getOptixVersion and guard in MakeHit test

MakeHit isn't functional in OptiX versions older than 9.0. Add a new
method, getOptixVersion, that returns the OptiX version of the current
device, or 0 if it's not supported (i.e. either no OptiX is found or the
device isn't CUDA).

* Formatting

* Revert "Formatting"

This reverts commit 7b79dc3ac37143445c86b3a90d6a42c76127bbc7.

* Revert "Add getOptixVersion and guard in MakeHit test"

This reverts commit 28d8aaf491892de69e5234fecea555ad20c43992.

* Guard MakeHit test based on OptiX version

---------

Co-authored-by: Simon Kallweit <64953474+skallweitNV@users.noreply.github.com>