Update slang-rhi for library rename and versioning (#559)
* Update for library rename and versioning

For shader-slang/slang#8828

Handle new filename patterns introduced by shader-slang/slang#4016

* Use only one slang-compiler check

* Move add_library(slang ..), etc, into a function

Avoid polluting the module namespace with temporary variables

* Rename function to hide name

Renaming the function to namespace it a bit, so it doesn't
accidentally conflict with one in a parent CMake file.

* Make filenames more explicit

Instead of building the filename in pieces, make the filename we're
checking for explicit for each platform.